+++ subsys/bluetooth/controller/hci/hci_internal.h
@@ -48,5 +48,3 @@ uint8_t hci_vendor_read_static_addr(struct bt_hci_vs_static_addr addrs[],
 void hci_vendor_read_key_hierarchy_roots(uint8_t ir[16], uint8_t er[16]);
 int hci_vendor_cmd_handle_common(uint16_t ocf, struct net_buf *cmd,
 			     struct net_buf **evt);
-void *hci_cmd_complete(struct net_buf **buf, uint8_t plen);
-void hci_evt_create(struct net_buf *buf, uint8_t evt, uint8_t len);
