+++ subsys/bluetooth/host/CMakeLists.txt
@@ -3,7 +3,7 @@
 zephyr_library()
 zephyr_library_link_libraries(subsys__bluetooth)
 
-zephyr_library_sources_ifdef(CONFIG_BT_HCI_RAW          hci_raw.c)
+zephyr_library_sources_ifdef(CONFIG_BT_HCI_RAW          hci_raw.c hci_common.c)
 zephyr_library_sources_ifdef(CONFIG_BT_DEBUG_MONITOR    monitor.c)
 zephyr_library_sources_ifdef(CONFIG_BT_TINYCRYPT_ECC    hci_ecc.c)
 zephyr_library_sources_ifdef(CONFIG_BT_A2DP             a2dp.c)
@@ -30,6 +30,7 @@ if(CONFIG_BT_HCI_HOST)
   zephyr_library_sources(
     uuid.c
     hci_core.c
+    hci_common.c
     )
   zephyr_library_sources_ifdef(
     CONFIG_BT_HOST_CRYPTO
